+++ src/mount/mount.spec.js
@@ -2,8 +2,14 @@ import 'angular';
 import 'angular-mocks';
 
 import mount from '.';
+import { createTemplate } from './template';
+import { compile } from './component';
+import { validate } from './validation';
 import TestElementWrapper from './TestElementWrapper';
 
+jest.mock('./template', () => ({ createTemplate: jest.fn() }));
+jest.mock('./component', () => ({ compile: jest.fn() }));
+jest.mock('./validation', () => ({ validate: jest.fn() }));
 jest.mock(
   './TestElementWrapper',
   () =>
@@ -13,44 +19,39 @@ jest.mock(
       }
     },
 );
-
 describe('mount', () => {
-  it('returns a test element wrapper created from angular element', () => {
-    initAngularComponents({
-      someComponent: {
-        template: '<div>Some content</div>',
-      },
-    });
+  afterEach(() => {
+    jest.clearAllMocks();
+  });
 
-    const component = mount('<some-component></some-component>');
+  it('validates by tag', () => {
+    expect(validate).not.toBeCalled();
+    mount('a-component');
+    expect(validate).toBeCalledWith('a-component');
+  });
+
+  it('creates template for tag', () => {
+    expect(createTemplate).not.toBeCalled();
+    mount('a-component');
+    expect(createTemplate).toBeCalledWith('a-component');
+  });
+
+  it('compiles created template with props', () => {
+    const props = { prop: 'Value' };
+    createTemplate.mockImplementation(tag => `template for ${tag}`);
+
+    expect(compile).not.toBeCalled();
+    mount('a-component', props);
+    expect(compile).toBeCalledWith('template for a-component', props);
+  });
+
+  it('returns test element wrapper created from angular element', () => {
+    const angularElement = { angular: 'element' };
+    compile.mockReturnValue(angularElement);
+
+    const component = mount('a-component');
 
     expect(component).toBeInstanceOf(TestElementWrapper);
-    expect(component.angularElementForSpec).toBeDefined();
-  });
-
-  it('makes props accessible through $ctrl', () => {
-    initAngularComponents({
-      someComponent: {
-        bindings: { someProp: '<' },
-        template: '<div>{{ $ctrl.someProp }}</div>',
-      },
-    });
-
-    const { angularElementForSpec } = mount(
-      '<some-component some-prop="$ctrl.someProp"></some-component>',
-      { someProp: 'Prop value' },
-    );
-
-    expect(angularElementForSpec.text()).toBe('Prop value');
+    expect(component.angularElementForSpec).toBe(angularElement);
   });
 });
-
-function initAngularComponents(components) {
-  const angularModule = angular.module('someModule', []);
-
-  Object.keys(components).forEach(name => {
-    angularModule.component(name, components[name]);
-  });
-
-  angular.mock.module('someModule');
-}