+++ src/mount/validation/AngularjsEnzymeError.js
@@ -1,0 +1,5 @@
+export default class AngularjsEnzymeError extends Error {
+  constructor(message) {
+    super(`AngularJS Enzyme Error: ${message}`);
+  }
+}