+++ src/mount/validation/AngularjsEnzymeError.spec.js
@@ -1,0 +1,9 @@
+import AngularjsEnzymeError from './AngularjsEnzymeError';
+
+describe('AngularJS Enzyme error', () => {
+  it('prepends error message with "AngularJS Enzyme Error: "', () => {
+    expect(new AngularjsEnzymeError('Some message').message).toBe(
+      'AngularJS Enzyme Error: Some message',
+    );
+  });
+});